remove DbtAssetsDefinition
This removes the DbtAssetsDefinition class. In its stead, it adds a get_asset_key_for_model function as a top-level export of dagster_dbt.

branch-name: no-dbt-assets-def
