+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/assets/dependencies.py
@@ -2,7 +2,13 @@ from typing import Any
 
 import pandas as pd
 from dagster import AssetOut, OpExecutionContext, Output, asset, multi_asset
-from dagster_dbt import DbtCliResource, dbt_assets
+from dagster_dbt import (
+    DbtCliResource,
+    dbt_assets,
+    get_asset_key_for_model,
+    get_asset_key_for_source,
+    get_asset_keys_by_output_name_for_source,
+)
 
 from ..constants import MANIFEST_PATH
 
@@ -13,7 +19,7 @@ def my_dbt_assets(context: OpExecutionContext, dbt: DbtCliResource):
 
 
 @asset(
-    key=my_dbt_assets.get_asset_key_for_source("raw_data"),
+    key=get_asset_key_for_source([my_dbt_assets], "raw_data"),
 )
 def raw_inventory() -> Any:
     return pd.DataFrame()
@@ -22,8 +28,8 @@ def raw_inventory() -> Any:
 @multi_asset(
     outs={
         name: AssetOut(key=asset_key)
-        for name, asset_key in my_dbt_assets.get_asset_keys_by_output_name_for_source(
-            "clients_data"
+        for name, asset_key in get_asset_keys_by_output_name_for_source(
+            [my_dbt_assets], "clients_data"
         ).items()
     }
 )
@@ -33,6 +39,6 @@ def clients_data(context):
     yield Output(value=pd.DataFrame(), output_name=output_names[1])
 
 
-@asset(deps=[my_dbt_assets.get_asset_key_for_model("customers")])
+@asset(deps=[get_asset_key_for_model([my_dbt_assets], "customers")])
 def cleaned_customers() -> Any:
     return pd.DataFrame()
