+++ examples/experimental/tutorial_dbt_dagster_v2/tutorial_dbt_dagster_v2/schedules/define_schedules.py
@@ -1,4 +1,4 @@
-from dagster_dbt import DbtCliResource, dbt_assets
+from dagster_dbt import DbtCliResource, build_schedule_from_dbt_selection, dbt_assets
 
 from ..constants import MANIFEST_PATH
 
@@ -8,13 +8,16 @@ def all_dbt_assets(context, dbt: DbtCliResource):
     yield from dbt.cli(["build"], context=context).stream()
 
 
-daily_dbt_assets_schedule = all_dbt_assets.build_schedule_from_dbt_selection(
+daily_dbt_assets_schedule = build_schedule_from_dbt_selection(
+    [all_dbt_assets],
     job_name="all_dbt_assets",
     cron_schedule="0 0 * * *",
     dbt_select="fqn:*",
 )
 
-hourly_staging_dbt_assets = all_dbt_assets.build_schedule_from_dbt_selection(
+
+hourly_staging_dbt_assets = build_schedule_from_dbt_selection(
+    [all_dbt_assets],
     job_name="staging_dbt_assets",
     cron_schedule="0 * * * *",
     dbt_select="fqn:staging.*",
