+++ python_modules/libraries/dagster-dbt/dagster_dbt_tests/core/test_asset_selection.py
@@ -5,6 +5,7 @@ from typing import Optional, Set
 import pytest
 from dagster._core.definitions.asset_graph import AssetGraph
 from dagster._core.definitions.events import AssetKey
+from dagster_dbt import build_dbt_asset_selection
 from dagster_dbt.asset_decorator import dbt_assets
 
 manifest_path = Path(__file__).parent.joinpath("..", "sample_manifest.json")
@@ -111,7 +112,8 @@ def test_dbt_asset_selection(
         ...
 
     asset_graph = AssetGraph.from_assets([my_dbt_assets])
-    asset_selection = my_dbt_assets.build_dbt_asset_selection(
+    asset_selection = build_dbt_asset_selection(
+        [my_dbt_assets],
         dbt_select=select or "fqn:*",
         dbt_exclude=exclude,
     )
